Some cleanup refactoring. Move Dense- and Sparse- TileData implementations to an impl package. Move the TileDataView object into the same package, and rename it SubTileDataView, in prep for more TileDataViews.
