+++ annotation-service/src/main/java/com/oculusinfo/annotation/AnnotationTile.java
@@ -24,9 +24,9 @@
 package com.oculusinfo.annotation;
 
 import com.oculusinfo.binning.BinIndex;
-import com.oculusinfo.binning.DenseTileData;
 import com.oculusinfo.binning.TileData;
 import com.oculusinfo.binning.TileIndex;
+import com.oculusinfo.binning.impl.DenseTileData;
 import com.oculusinfo.factory.util.Pair;
 
 import java.util.ArrayList;
