+++ annotation-service/src/main/java/com/oculusinfo/annotation/util/AnnotationUtil.java
@@ -29,8 +29,8 @@ import com.oculusinfo.annotation.AnnotationData;
 import com.oculusinfo.annotation.AnnotationTile;
 import com.oculusinfo.annotation.index.AnnotationIndexer;
 import com.oculusinfo.binning.BinIndex;
-import com.oculusinfo.binning.DenseTileData;
 import com.oculusinfo.binning.TileIndex;
+import com.oculusinfo.binning.impl.DenseTileData;
 import com.oculusinfo.factory.util.Pair;
 import org.json.JSONArray;
 import org.json.JSONObject;
