+++ binning-utilities/src/main/java/com/oculusinfo/binning/io/impl/DummyPyramidIO.java
@@ -32,9 +32,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
 
-import com.oculusinfo.binning.DenseTileData;
 import com.oculusinfo.binning.TileData;
 import com.oculusinfo.binning.TileIndex;
+import com.oculusinfo.binning.impl.DenseTileData;
 import com.oculusinfo.binning.io.PyramidIO;
 import com.oculusinfo.binning.io.serialization.TileSerializer;
