+++ binning-utilities/src/test/java/com/oculusinfo/binning/io/impl/FileSystemPyramidSourceTest.java
@@ -25,8 +25,8 @@
 package com.oculusinfo.binning.io.impl;
 
 import com.oculusinfo.binning.TileData;
-import com.oculusinfo.binning.DenseTileData;
 import com.oculusinfo.binning.TileIndex;
+import com.oculusinfo.binning.impl.DenseTileData;
 import com.oculusinfo.binning.io.PyramidIO;
 import com.oculusinfo.binning.io.serialization.TileSerializer;
 import com.oculusinfo.binning.io.serialization.impl.PrimitiveAvroSerializer;
