+++ binning-utilities/src/test/java/com/oculusinfo/binning/io/serialization/impl/PairArrayAvroSerializerTests.java
@@ -34,9 +34,9 @@ import org.apache.avro.file.CodecFactory;
 import org.junit.Assert;
 import org.junit.Test;
 
-import com.oculusinfo.binning.DenseTileData;
 import com.oculusinfo.binning.TileData;
 import com.oculusinfo.binning.TileIndex;
+import com.oculusinfo.binning.impl.DenseTileData;
 import com.oculusinfo.binning.io.serialization.TileSerializer;
 import com.oculusinfo.factory.util.Pair;
