+++ binning-utilities/src/test/java/com/oculusinfo/binning/io/serialization/impl/PrimitiveArrayAvroSerializerTest.java
@@ -35,8 +35,8 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.oculusinfo.binning.TileData;
-import com.oculusinfo.binning.DenseTileData;
 import com.oculusinfo.binning.TileIndex;
+import com.oculusinfo.binning.impl.DenseTileData;
 import com.oculusinfo.binning.io.serialization.TileSerializer;
 
 public class PrimitiveArrayAvroSerializerTest {
