+++ tile-rendering/src/main/java/com/oculusinfo/tile/rendering/transformations/tile/FilterVarsDoubleArrayTileTransformer.java
@@ -28,9 +28,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.oculusinfo.binning.TileData;
-import com.oculusinfo.binning.DenseTileData;
-
 import com.oculusinfo.binning.TileIndex;
+import com.oculusinfo.binning.impl.DenseTileData;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.json.JSONArray;
